implements G4-92 Created Tests for all methods of class Cake and testing method getAllUsersByRole in class User
