+++ src/main/resources/sql/cakes.sql
@@ -1,4 +1,4 @@
-INSERT INTO cakes(CATEGORY, INGREDIENTS, NAME, PRICE)
-VALUES ('CHEESECAKES', 'egg,milk,salt', 'vanilla-cheesecake', 150);
-INSERT INTO cakes(CATEGORY, INGREDIENTS, NAME, PRICE)
-VALUES ('CUPCAKES', 'milk,egg,salt', 'blueberry-cupcake', 100);
+INSERT INTO cakes(CATEGORY, STATE, INGREDIENTS, NAME, PRICE)
+VALUES ('CHEESECAKES', 'CREATED', 'egg,milk,salt', 'vanilla-cheesecake', 150);
+INSERT INTO cakes(CATEGORY, STATE, INGREDIENTS, NAME, PRICE)
+VALUES ('CUPCAKES', 'CREATED', 'milk,egg,salt', 'blueberry-cupcake', 100);
