+++ src/main/resources/sql/users.sql
@@ -1,0 +1,8 @@
+INSERT INTO PUBLIC.ACCOUNT (EMAIL, FULL_NAME, HASH_PASSWORD, HOUSE_NUMBER, PHONE_NUMBER, ROLE, STATE, STREET, TOWN)
+VALUES ('konditerow@gmail.com', 'Konditer Konditerow', '$2a$10$XhZqBs2ID5aIey7WJJiPAexSGXcfuB6NFSb/ZC/S3GMkO/ouhmTdG', 78, '+4917688359755', 'CONFECTIONER', 'CONFIRMED', 'StrandStrasse', 'Kiel');
+INSERT INTO PUBLIC.ACCOUNT (EMAIL, FULL_NAME, HASH_PASSWORD, HOUSE_NUMBER, PHONE_NUMBER, ROLE, STATE, STREET, TOWN)
+VALUES ('konditeryan@gmail.com', 'Konditer Konditeryan', '$2a$10$jtLDeFSt7C1jNGtxpzh6kurTi.rej21OBRqjMexaC9jbW4JI967bC', 12, '+4917612359755', 'CONFECTIONER', 'CONFIRMED', 'StrandStrasse', 'Kiel');
+INSERT INTO PUBLIC.ACCOUNT (EMAIL, FULL_NAME, HASH_PASSWORD, HOUSE_NUMBER, PHONE_NUMBER, ROLE, STATE, STREET, TOWN)
+VALUES ('konditeridze@gmail.com', 'Konditer Konditeridze', '$2a$10$KrckCT.RKuAbniEAJoi2R.xIoeTHin/PEjuPH4qRVtTV9DStncu1K', 99, '+4917333359755', 'CONFECTIONER', 'CONFIRMED', 'StrandStrasse', 'Kiel');
+INSERT INTO PUBLIC.ACCOUNT (EMAIL, FULL_NAME, HASH_PASSWORD, HOUSE_NUMBER, PHONE_NUMBER, ROLE, STATE, STREET, TOWN)
+VALUES ('konditerenko@gmail.com', 'Konditer Konditerenko', '$2a$10$ruOLOFb6ugIquDl4iSv0DOVfSpC1orM5E512eitSFPoaJF03EfhP2', 19, '+4917612359755', 'CONFECTIONER', 'CONFIRMED', 'StrandStrasse', 'Kiel');
