+++ src/test/java/de/ait/artcake/controllers/UserControllerIntegrationTest.java
@@ -1,0 +1,41 @@
+package de.ait.artcake.controllers;
+
+import org.junit.jupiter.api.*;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
+import org.springframework.security.test.context.support.WithUserDetails;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.jdbc.Sql;
+import org.springframework.test.web.servlet.MockMvc;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+@DisplayNameGeneration(value = DisplayNameGenerator.ReplaceUnderscores.class)
+@DisplayName("UsersController is works: ")
+@ActiveProfiles("test")
+public class UserControllerIntegrationTest {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Nested
+    @DisplayName("GET /api/users/role/{role} method is works: ")
+    class GetUsersByRoleTests {
+
+        @WithUserDetails(value = "test@mail.com")
+        @Sql(scripts = "/sql/data_for_users.sql")
+        @DirtiesContext(methodMode = DirtiesContext.MethodMode.AFTER_METHOD)
+        @Test
+        public void getAllUsersByRole() throws Exception {
+            mockMvc.perform(get("/api/users/role/CONFECTIONER")
+                            .contentType(MediaType.APPLICATION_JSON))
+                    .andExpect(status().isOk());
+        }
+    }
+}
