+++ src/test/resources/sql/data_for_users.sql
@@ -1,0 +1,9 @@
+insert into account(full_name, email, hash_password,house_number, phone_number, role, state, street, town)
+values ('Jim Carrey','test@mail.com','$2a$10$pzso4fXl0Lod4xofbi49Ueoa7aFO3KgNQbzAYW4iJ3zCPgcGdLvkm','55', '+4917611223344', 'MANAGER', 'CONFIRMED','Norbert-Schmid-Platz 11', 'Hamburg');
+
+
+INSERT INTO PUBLIC.ACCOUNT (EMAIL, FULL_NAME, HASH_PASSWORD, HOUSE_NUMBER, PHONE_NUMBER, ROLE, STATE, STREET, TOWN)
+VALUES ('konditerow@gmail.com', 'Konditer Konditerow', '$2a$10$XhZqBs2ID5aIey7WJJiPAexSGXcfuB6NFSb/ZC/S3GMkO/ouhmTdG', 78, '+4917688359755', 'CONFECTIONER', 'CONFIRMED', 'StrandStrasse', 'Kiel');
+INSERT INTO PUBLIC.ACCOUNT (EMAIL, FULL_NAME, HASH_PASSWORD, HOUSE_NUMBER, PHONE_NUMBER, ROLE, STATE, STREET, TOWN)
+VALUES ('konditeryan@gmail.com', 'Konditer Konditeryan', '$2a$10$jtLDeFSt7C1jNGtxpzh6kurTi.rej21OBRqjMexaC9jbW4JI967bC', 12, '+4917612359755', 'CONFECTIONER', 'CONFIRMED', 'StrandStrasse', 'Kiel');
+
